Moved core functions from RDDLLiftedModel to RDDLModel
